customer-mcp-container/pyproject.toml
[project]
name = "customer"
version = "0.1.0"
description = "MCP server to provide customer interactions"
authors = [{name = "Ben Greene",email = "BenJGreene@gmail.com"}]
readme = "README.md"
requires-python = "^3.12"

[tool.poetry.dependencies]

click =  "^8.3"
prometheus-async = "^25.1"
pydantic-settings = {extras = ["yaml"], version = "^2.12"}
fastmcp = "^2.14"
fastapi = "^0.128"



[tool.poetry.group.dev.dependencies]
aiohttp-devtools = "^1"
pytest = "^8"
Sphinx = "^9"
sphinx-autodoc-typehints = "^3.6"
pytest-sphinx = "^0.6"
flake8 = "^7"
flake8-black = "^0.4"
pylint = "^4.0"
sphinx-click = "^6.0.0"
pytest-click = "^1.1.0"
pytest-cov = "^7.0"
pytest-watcher = "^0.6"
pytest-aiohttp = "^1.1.0"


[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.pyrefly]
# set the directory Pyrefly will search for files to type check
project-includes = [
    "customer"
]

# manually set the `sys.platform` Pyrefly will assume when type checking
python-platform = "linux"

# a table mapping error codes to an `is-enabled` boolean
[errors]
# disable `bad-assignment` errors
bad-assignment = false
# disable `bad-return` errors
bad-return = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
customer = "customer.cli:cli"
